separate nophone component and improve check
